Move a lot of code into separate files
