+++ vsdbg-engine-extension/.clang-tidy
@@ -32,7 +32,7 @@ Checks: >
   -cppcoreguidelines-owning-memory,
   -cppcoreguidelines-pro-type-union-access,
 FormatStyle: file
-HeaderFilterRegex: ^.*\\src\\.*\.h$
+HeaderFilterRegex: ^.*\\src\\[a-zA-Z0-9]*\.h(pp)?$
 CheckOptions:
   - key: readability-identifier-naming.ClassCase
     value: CamelCase
