+++ vsdbg-engine-extension/src/ChildDebuggerService.cpp
@@ -1,25 +1,27 @@
+
+#include "ChildDebuggerService.hpp"
+
 #include <cassert>
 #include <chrono>
 #include <filesystem>
+#include <iostream>
 #include <memory>
 #include <string>
 #include <thread>
 
-#include <iostream> // TODO: remove
-
-#include <atlbase.h>
-#include <atlcom.h>
-#include <atlctl.h>
-
 #include <psapi.h>
-#include <tlhelp32.h>
 
 #include <vsdebugeng.h>
 #include <vsdebugeng.templates.h>
 
-#include "ChildDebuggerService.h"
-
 #include "nlohmann/json.hpp"
+
+#include "CreateFunctionFrames.hpp"
+#include "CreateFunctionType.hpp"
+#include "CustomMessageType.hpp"
+#include "DataItems.hpp"
+#include "SystemUtils.hpp"
+#include "UnicodeUtils.hpp"
 
 using namespace Microsoft::VisualStudio::Debugger;
 
@@ -30,445 +32,21 @@ static const GUID source_id = {
     0x0bb8'9d05, 0x9ead, 0x4295, {0x9a, 0x74, 0xa2, 0x41, 0x58, 0x3d, 0xe4, 0x20}
 };
 
-// Takes the Wide-Char (UTF-16) encoded input string and returns it as an UTF-8 encoded string.
-std::string utf8_encode(const std::wstring& input)
-{
-    if(input.empty()) return {};
-
-    const auto result_size = WideCharToMultiByte(CP_UTF8, 0,
-                                                 input.data(), static_cast<int>(input.size()),
-                                                 nullptr, 0,
-                                                 nullptr, nullptr);
-    assert(result_size > 0);
-
-    std::string result(result_size, '\0');
-    const auto  bytes_converted = WideCharToMultiByte(CP_UTF8, 0,
-                                                      input.data(), static_cast<int>(input.size()),
-                                                      result.data(), result_size,
-                                                      nullptr, nullptr);
-    assert(bytes_converted != 0);
-
-    return result;
-}
-// Takes the UTF-8 encoded input string and returns it as an Wide-Char (UTF-16) encoded string.
-std::wstring utf16_encode(const std::string& input)
-{
-    if(input.empty()) return {};
-
-    const auto result_size = MultiByteToWideChar(CP_UTF8, 0,
-                                                 input.data(), static_cast<int>(input.size()),
-                                                 nullptr, 0);
-    assert(result_size > 0);
-
-    std::wstring result(result_size, L'\0');
-    const auto   bytes_converted = MultiByteToWideChar(CP_UTF8, 0,
-                                                       input.data(), static_cast<int>(input.size()),
-                                                       result.data(), result_size);
-    assert(bytes_converted != 0);
-
-    return result;
-}
-
-// Retrieve the path to the module (DLL) that holds this function.
-std::optional<std::filesystem::path> get_current_module_path()
-{
-    HMODULE           current_module          = nullptr;
-    const auto* const module_function_address = (LPCWSTR)(&get_current_module_path); // NOLINT(cppcoreguidelines-pro-type-cstyle-cast)
-    if(GetModuleHandleExW(GET_MODULE_HANDLE_EX_FLAG_FROM_ADDRESS | GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT, module_function_address, &current_module) == FALSE) return std::nullopt;
-
-    std::array<WCHAR, MAX_PATH> buffer; // FIXME: handle paths longer than MAX_PATH
-
-    const auto result_size = GetModuleFileNameW(current_module, buffer.data(), buffer.size());
-    if(result_size == 0) return std::nullopt;
-
-    auto result = std::filesystem::path(buffer.data()).parent_path();
-    if(result.empty()) return std::nullopt;
-
-    return result;
-}
-
-enum class CreateFunctionType
-{
-    create_process,
-    create_process_as_user,
-    create_process_with_token,
-    create_process_with_logon
-};
-
-enum class CustomMessageType
-{
-    settings      = 1,
-    resume_child  = 2,
-    resume_parent = 3,
-    inform_child  = 4,
-};
-
-// Base class for breakpoint information classes.
-template<typename Interface>
-class BaseObject : public Interface
-{
-    volatile LONG ref_count_;
-
-public:
-    virtual ~BaseObject() = default;
-
-    BaseObject() = default;
-
-    BaseObject(BaseObject&)             = delete;
-    BaseObject(BaseObject&&)            = delete;
-    BaseObject& operator=(BaseObject&)  = delete;
-    BaseObject& operator=(BaseObject&&) = delete;
-
-    ULONG __stdcall AddRef() override
-    {
-        return (ULONG)InterlockedIncrement(&ref_count_);
-    }
-    ULONG __stdcall Release() override
-    {
-        auto result = (ULONG)InterlockedDecrement(&ref_count_);
-        if(result == 0)
-        {
-            delete this;
-        }
-        return result;
-    }
-    HRESULT __stdcall QueryInterface(REFIID riid, _Deref_out_ void** ppv) override
-    {
-        if(riid == __uuidof(IUnknown))
-        {
-            *ppv = static_cast<IUnknown*>(this);
-            AddRef();
-            return S_OK;
-        }
-        // This example is implementing the optional interface IDkmDisposableDataItem
-        if(riid == __uuidof(IDkmDisposableDataItem))
-        {
-            *ppv = static_cast<IDkmDisposableDataItem*>(this);
-            AddRef();
-            return S_OK;
-        }
-
-        *ppv = nullptr;
-        return E_NOINTERFACE;
-    }
-};
-
-// Class holding additional information for breakpoints to be triggered on
-// a call to any of the process creation functions.
-class __declspec(uuid("{1483C347-BDAD-4626-B33F-D16970542239}")) CreateInInfo :
-    public BaseObject<IDkmDisposableDataItem>
-{
-    bool               isUnicode_;
-    CreateFunctionType functionType_;
-
-public:
-    explicit CreateInInfo(bool is_unicode, CreateFunctionType function_type) :
-        isUnicode_(is_unicode),
-        functionType_(function_type)
-    {}
-
-    CreateInInfo(CreateInInfo&)             = delete;
-    CreateInInfo(CreateInInfo&&)            = delete;
-    CreateInInfo& operator=(CreateInInfo&)  = delete;
-    CreateInInfo& operator=(CreateInInfo&&) = delete;
-
-    ~CreateInInfo() override = default;
-
-    HRESULT __stdcall OnClose() override
-    {
-        return S_OK;
-    }
-
-    [[nodiscard]] bool get_is_unicode() const
-    {
-        return isUnicode_;
-    }
-
-    [[nodiscard]] CreateFunctionType get_function_type() const
-    {
-        return functionType_;
-    }
-};
-
-// Class holding additional information for breakpoints to be triggered when
-// we return from a  call to any of the process creation functions.
-class __declspec(uuid("{F1AB4299-C3EB-47C5-83B7-813E28B9DA89}")) CreateOutInfo :
-    public BaseObject<IDkmDisposableDataItem>
-{
-    UINT64 lpProcessInformation_;
-    bool   suspended_;
-
-public:
-    explicit CreateOutInfo(UINT64 process_information, bool suspended) :
-        lpProcessInformation_(process_information),
-        suspended_(suspended)
-    {}
-
-    CreateOutInfo(CreateOutInfo&)             = delete;
-    CreateOutInfo(CreateOutInfo&&)            = delete;
-    CreateOutInfo& operator=(CreateOutInfo&)  = delete;
-    CreateOutInfo& operator=(CreateOutInfo&&) = delete;
-
-    ~CreateOutInfo() override = default;
-
-    HRESULT __stdcall OnClose() override
-    {
-        return S_OK;
-    }
-
-    [[nodiscard]] UINT64 get_process_information_address() const
-    {
-        return lpProcessInformation_;
-    }
-
-    [[nodiscard]] bool get_suspended() const
-    {
-        return suspended_;
-    }
-};
-
-class ATL_NO_VTABLE __declspec(uuid("{0709D0FC-76B1-44E8-B781-E8C43461CFAC}")) ChildProcessDataItem :
-    public IUnknown,
-    public CComObjectRootEx<CComMultiThreadModel>
-{
-    bool passedInitialBreakpoint_{false};
-
-public:
-    explicit ChildProcessDataItem() = default;
-
-    ~ChildProcessDataItem() = default;
-
-    ChildProcessDataItem(ChildProcessDataItem&)             = delete;
-    ChildProcessDataItem(ChildProcessDataItem&&)            = delete;
-    ChildProcessDataItem& operator=(ChildProcessDataItem&)  = delete;
-    ChildProcessDataItem& operator=(ChildProcessDataItem&&) = delete;
-
-    [[nodiscard]] bool get_passed_initial_breakpoint() const
-    {
-        return passedInitialBreakpoint_;
-    }
-
-    void set_passed_initial_breakpoint()
-    {
-        passedInitialBreakpoint_ = true;
-    }
-
-protected:
-    // NOLINTNEXTLINE(bugprone-reserved-identifier, readability-identifier-naming)
-    HRESULT _InternalQueryInterface(REFIID riid, void** object)
-    {
-        if(object == nullptr)
-            return E_POINTER;
-
-        if(riid == __uuidof(IUnknown))
-        {
-            *object = static_cast<IUnknown*>(this);
-            AddRef();
-            return S_OK;
-        }
-
-        *object = nullptr;
-        return E_NOINTERFACE;
-    }
-};
-
-struct CreateProcessStack
-{
-#if defined(_X86_) || defined(_ARM_)
-    using ptr_int_t = UINT32;
-#else
-    using ptr_int_t = UINT64;
-#endif
-
-    ptr_int_t return_address;
-
-    ptr_int_t lpApplicationName;   // NOLINT(readability-identifier-naming)
-    ptr_int_t lpCommandLine;       // NOLINT(readability-identifier-naming)
-    ptr_int_t lpProcessAttributes; // NOLINT(readability-identifier-naming)
-    ptr_int_t lpThreadAttributes;  // NOLINT(readability-identifier-naming)
-    UINT8     bInheritHandles;     // NOLINT(readability-identifier-naming)
-    UINT8     Padding1;            // NOLINT(readability-identifier-naming)
-    UINT16    Padding2;            // NOLINT(readability-identifier-naming)
-#if defined(_AMD64_) || defined(_ARM64_)
-    UINT32 Padding3; // NOLINT(readability-identifier-naming)
-#endif
-    UINT32 dwCreationFlags; // NOLINT(readability-identifier-naming)
-#if defined(_AMD64_) || defined(_ARM64_)
-    // UINT32       Padding4;             // NOLINT(readability-identifier-naming)
-#endif
-    ptr_int_t lpEnvironment;        // NOLINT(readability-identifier-naming)
-    ptr_int_t lpCurrentDirectory;   // NOLINT(readability-identifier-naming)
-    ptr_int_t lpStartupInfo;        // NOLINT(readability-identifier-naming)
-    ptr_int_t lpProcessInformation; // NOLINT(readability-identifier-naming)
-
-    // NOLINTNEXTLINE(readability-identifier-naming, readability-convert-member-functions-to-static)
-    [[nodiscard]] ptr_int_t get_lpApplicationName([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_)
-        return lpApplicationName;
-#elif defined(_AMD64_)
-        return context.Rcx;
-#elif defined(_ARM_)
-        return context.R9;
-#elif defined(_ARM64_)
-        return context.X9;
-#endif
-    }
-
-    // NOLINTNEXTLINE(readability-identifier-naming, readability-convert-member-functions-to-static)
-    [[nodiscard]] ptr_int_t get_lpCommandLine([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_)
-        return lpCommandLine;
-#elif defined(_AMD64_)
-        return context.Rdx;
-#elif defined(_ARM_)
-        return context.R1;
-#elif defined(_ARM64_)
-        return context.X0;
-#endif
-    }
-
-    // NOLINTNEXTLINE(readability-identifier-naming, readability-convert-member-functions-to-static)
-    [[nodiscard]] DWORD64 get_lpProcessInformation([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_)
-        return lpProcessInformation;
-#elif defined(_AMD64_)
-        return lpProcessInformation;
-#elif defined(_ARM_)
-        return context.R2;
-#elif defined(_ARM64_)
-        return context.X2;
-#endif
-    }
-
-    // NOLINTNEXTLINE(readability-convert-member-functions-to-static)
-    [[nodiscard]] DWORD64 get_return_address([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_) || defined(_AMD64_)
-        return return_address;
-#elif defined(_ARM_)
-        return context.Lr;
-#elif defined(_ARM64_)
-        return context.Lr;
-#endif
-    }
-};
-
-struct CreateProcessAsUserStack
-{
-#if defined(_X86_) || defined(_ARM_)
-    using ptr_int_t = UINT32;
-#else
-    using ptr_int_t = UINT64;
-#endif
-
-    ptr_int_t return_address; // NOLINT(readability-identifier-naming)
-
-    ptr_int_t hToken; // NOLINT(readability-identifier-naming)
-
-    ptr_int_t lpApplicationName;   // NOLINT(readability-identifier-naming)
-    ptr_int_t lpCommandLine;       // NOLINT(readability-identifier-naming)
-    ptr_int_t lpProcessAttributes; // NOLINT(readability-identifier-naming)
-    ptr_int_t lpThreadAttributes;  // NOLINT(readability-identifier-naming)
-    UINT8     bInheritHandles;     // NOLINT(readability-identifier-naming)
-    UINT8     Padding1;            // NOLINT(readability-identifier-naming)
-    UINT16    Padding2;            // NOLINT(readability-identifier-naming)
-#if defined(_AMD64_) || defined(_ARM64_)
-    UINT32 Padding3; // NOLINT(readability-identifier-naming)
-#endif
-    UINT32 dwCreationFlags; // NOLINT(readability-identifier-naming)
-#if defined(_AMD64_) || defined(_ARM64_)
-    UINT32 Padding4; // NOLINT(readability-identifier-naming)
-#endif
-    ptr_int_t lpEnvironment;        // NOLINT(readability-identifier-naming)
-    ptr_int_t lpCurrentDirectory;   // NOLINT(readability-identifier-naming)
-    ptr_int_t lpStartupInfo;        // NOLINT(readability-identifier-naming)
-    ptr_int_t lpProcessInformation; // NOLINT(readability-identifier-naming)
-
-    // NOLINTNEXTLINE(readability-identifier-naming, readability-convert-member-functions-to-static)
-    [[nodiscard]] DWORD64 get_lpApplicationName([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_)
-        return lpApplicationName;
-#elif defined(_AMD64_)
-        return context.Rdx;
-#elif defined(_ARM_)
-        return context.R1;
-#elif defined(_ARM64_)
-        return context.X1;
-#endif
-    }
-
-    // NOLINTNEXTLINE(readability-identifier-naming, readability-convert-member-functions-to-static)
-    [[nodiscard]] DWORD64 get_lpCommandLine([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_)
-        return lpCommandLine;
-#elif defined(_AMD64_)
-        return context.R8;
-#elif defined(_ARM_)
-        return context.R2;
-#elif defined(_ARM64_)
-        return context.X2;
-#endif
-    }
-
-    // NOLINTNEXTLINE(readability-identifier-naming, readability-convert-member-functions-to-static)
-    [[nodiscard]] DWORD64 get_lpProcessInformation([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_)
-        return lpProcessInformation;
-#elif defined(_AMD64_)
-        return lpProcessInformation;
-#elif defined(_ARM_)
-        return lpProcessInformation;
-#elif defined(_ARM64_)
-        return lpProcessInformation;
-#endif
-    }
-
-    // NOLINTNEXTLINE(readability-convert-member-functions-to-static)
-    [[nodiscard]] DWORD64 get_return_address([[maybe_unused]] const CONTEXT& context) const
-    {
-#if defined(_X86_) || defined(_AMD64_)
-        return return_address;
-#elif defined(_ARM_)
-        return context.Lr;
-#elif defined(_ARM64_)
-        return context.Lr;
-#endif
-    }
-};
-
-[[nodiscard]] inline auto get_stack_pointer(const CONTEXT& context)
-{
-#if defined(_X86_)
-    return context.Esp;
-#elif defined(_AMD64_)
-    return context.Rsp;
-#elif defined(_ARM_)
-    return context.Sp;
-#elif defined(_ARM64_)
-    return context.Sp;
-#endif
-}
-
-[[nodiscard]] inline auto get_return_value(const CONTEXT& context)
-{
-#if defined(_X86_)
-    return context.Eax;
-#elif defined(_AMD64_)
-    return context.Rax;
-#elif defined(_ARM_)
-    return context.R0;
-#elif defined(_ARM64_)
-    return context.X0;
-#endif
-}
-
 // TODO: add parameter stack definitions for `CreateProcessWithToken` and `CreateProcessWithLogon`.
+
+template<typename T>
+T try_get_or(const nlohmann::json& json, std::string_view name, T default_value)
+{
+    if(json.count(name) == 0) return default_value;
+    return json.at(name).get<T>();
+}
+
+std::optional<std::wstring> try_get_optional_string(const nlohmann::json& json, std::string_view name)
+{
+    if(json.count(name) == 0) return std::nullopt;
+    const auto str = json.at(name).get<std::string>();
+    return utf8_to_utf16(str);
+}
 
 HRESULT read_string_from_memory_at(
     DkmProcess*         process,
@@ -504,8 +82,8 @@ bool check_attach_to_process(
     {
         log_file << "  Check process config: "
                  << "\n";
-        log_file << "    applicationName: " << (config.application_name ? utf8_encode(*config.application_name) : "<EMPTY>") << "\n";
-        log_file << "    commandLine: " << (config.command_line ? utf8_encode(*config.command_line) : "<EMPTY>") << "\n";
+        log_file << "    applicationName: " << (config.application_name ? utf16_to_utf8(*config.application_name) : "<EMPTY>") << "\n";
+        log_file << "    commandLine: " << (config.command_line ? utf16_to_utf8(*config.command_line) : "<EMPTY>") << "\n";
 
         // Skip invalid, empty config
         if(!config.application_name && !config.command_line) continue;
@@ -579,7 +157,7 @@ HRESULT handle_call_to_create_process(
     }
     if(application_name)
     {
-        log_file << "  APP " << utf8_encode(application_name->Value()) << "\n";
+        log_file << "  APP " << utf16_to_utf8(application_name->Value()) << "\n";
         log_file.flush();
     }
 
@@ -595,7 +173,7 @@ HRESULT handle_call_to_create_process(
     }
     if(command_line)
     {
-        log_file << "  CL " << utf8_encode(command_line->Value()) << "\n";
+        log_file << "  CL " << utf16_to_utf8(command_line->Value()) << "\n";
         log_file.flush();
     }
 
@@ -673,23 +251,6 @@ HRESULT handle_call_to_create_process(
     return S_OK;
 }
 
-std::optional<std::vector<std::string>> read_no_suspend(const std::filesystem::path& no_suspend_file_path)
-{
-    std::ifstream no_suspend_file(no_suspend_file_path);
-
-    if(!no_suspend_file.is_open()) return std::nullopt;
-
-    std::vector<std::string> result;
-    std::string              line;
-    while(std::getline(no_suspend_file, line))
-    {
-        if(line.empty()) continue;
-        if(line.starts_with('#')) continue;
-        result.push_back(line);
-    }
-    return result;
-}
-
 CChildDebuggerService::CChildDebuggerService()
 {
     const auto* const log_file_name = "ChildDebugger.log";
@@ -705,20 +266,6 @@ CChildDebuggerService::CChildDebuggerService()
     DkmString::Create(L"CreateProcessAsUserA", &create_process_function_names_[3]);
     DkmString::Create(L"CreateProcessWithTokenW", &create_process_function_names_[4]);
     DkmString::Create(L"CreateProcessWithLogonW", &create_process_function_names_[5]);
-}
-
-template<typename T>
-T try_get_or(const nlohmann::json& json, std::string_view name, T default_value)
-{
-    if(json.count(name) == 0) return default_value;
-    return json.at(name).get<T>();
-}
-
-std::optional<std::wstring> try_get_optional_string(const nlohmann::json& json, std::string_view name)
-{
-    if(json.count(name) == 0) return std::nullopt;
-    const auto str = json.at(name).get<std::string>();
-    return utf16_encode(str);
 }
 
 HRESULT STDMETHODCALLTYPE CChildDebuggerService::SendLower(
@@ -739,7 +286,7 @@ HRESULT STDMETHODCALLTYPE CChildDebuggerService::SendLower(
 
         // TODO: if we can find a reasonably small JSON parser library that can handle UTF-16 encoded
         // strings, we can probably get rid of this transcoding.
-        const auto utf8_settings_str = utf8_encode(settings_str);
+        const auto utf8_settings_str = utf16_to_utf8(settings_str);
 
         try
         {
@@ -777,8 +324,8 @@ HRESULT STDMETHODCALLTYPE CChildDebuggerService::SendLower(
         log_file_ << "  processConfigs:\n";
         for(const auto& config : settings_.process_configs)
         {
-            log_file_ << "    applicationName:        " << (config.application_name ? utf8_encode(*config.application_name) : "<EMPTY>") << "\n";
-            log_file_ << "    commandLine:            " << (config.command_line ? utf8_encode(*config.command_line) : "<EMPTY>") << "\n";
+            log_file_ << "    applicationName:        " << (config.application_name ? utf16_to_utf8(*config.application_name) : "<EMPTY>") << "\n";
+            log_file_ << "    commandLine:            " << (config.command_line ? utf16_to_utf8(*config.command_line) : "<EMPTY>") << "\n";
             log_file_ << "    attach:                 " << config.attach << "\n";
         }
         log_file_.flush();
@@ -919,9 +466,9 @@ HRESULT STDMETHODCALLTYPE CChildDebuggerService::OnModuleInstanceLoad(
         if(native_module_instance->FindExportName(function_name, true, &address) != S_OK) continue;
 
         log_file_ << "OnModuleInstanceLoad (Debugger PID " << GetCurrentProcessId() << ")\n";
-        log_file_ << "  " << utf8_encode(module_instance->Name()->Value()) << "\n";
+        log_file_ << "  " << utf16_to_utf8(module_instance->Name()->Value()) << "\n";
         log_file_ << "  Base address " << module_instance->BaseAddress() << "\n";
-        log_file_ << "  Function address " << utf8_encode(function_name->Value()) << " @" << address->RVA() << "\n";
+        log_file_ << "  Function address " << utf16_to_utf8(function_name->Value()) << " @" << address->RVA() << "\n";
         log_file_.flush();
 
         // TODO: Simplify this:
@@ -975,7 +522,7 @@ HRESULT STDMETHODCALLTYPE CChildDebuggerService::OnRuntimeBreakpoint(
     log_file_ << "OnRuntimeBreakpoint (Debugger PID " << GetCurrentProcessId() << ")\n";
     std::array<wchar_t, 64> guid_str = {L'\0'};
     StringFromGUID2(runtime_breakpoint->SourceId(), guid_str.data(), guid_str.size());
-    log_file_ << "  Source ID:" << utf8_encode(guid_str.data()) << "\n";
+    log_file_ << "  Source ID:" << utf16_to_utf8(guid_str.data()) << "\n";
     log_file_.flush();
 
     CComPtr<CreateInInfo> in_info;
@@ -1076,7 +623,7 @@ HRESULT STDMETHODCALLTYPE CChildDebuggerService::OnRuntimeBreakpoint(
             CloseHandle(process_handle);
         }
 
-        log_file_ << "  Child App Name " << utf8_encode(application_name) << "\n";
+        log_file_ << "  Child App Name " << utf16_to_utf8(application_name) << "\n";
         log_file_ << "  Child PID " << proc_info.dwProcessId << " TID " << proc_info.dwThreadId << "\n";
         log_file_ << "  Child P-HANDLE " << proc_info.hProcess << " T-HANDLE " << proc_info.hThread << "\n";
         log_file_.flush();
