+++ vsdbg-engine-extension/src/CreateFunctionType.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+enum class CreateFunctionType
+{
+    create_process,
+    create_process_as_user,
+    create_process_with_token,
+    create_process_with_logon
+};
