+++ vsdbg-engine-extension/src/CustomMessageType.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+enum class CustomMessageType
+{
+    settings      = 1,
+    resume_child  = 2,
+    resume_parent = 3,
+    inform_child  = 4,
+};
